+++ src/Tests/AuthorizationT/ScopedRightsTests.cs
@@ -15,14 +15,14 @@ namespace Tests.AuthorizationT
 
         public GlobalScopedRightsTests()
         {
-            _sut = new ScopedRights(ScopeId.Global, new short[] { 2, 5});
+            _sut = new ScopedRights(ScopeId.Global, new int[] { 2, 5});
         }
 
         [Theory]
-        [InlineData((short)2,(short)5)]
-        [InlineData((short)2,(short)7)]
-        [InlineData((short)9,(short)5)]
-        public void existing_rights_pass_check(short right1,short right2)
+        [InlineData((int)2,(int)5)]
+        [InlineData((int)2,(int)7)]
+        [InlineData((int)9,(int)5)]
+        public void existing_rights_pass_check(int right1,int right2)
         {
             _sut.HasRight(right1, right2).Should().BeTrue();
         }
@@ -44,7 +44,7 @@ namespace Tests.AuthorizationT
         [Fact]
         public void when_has_scoped_admin_check_fails()
         {
-            var sut = new ScopedRights(ScopeId.Global, new short[] {ScopedRights.ScopedAdmin});
+            var sut = new ScopedRights(ScopeId.Global, new int[] {ScopedRights.ScopedAdmin});
             sut.HasRight(23).Should().BeFalse();
         }
 
@@ -72,7 +72,7 @@ namespace Tests.AuthorizationT
         [Fact]
         public void when_has_global_admin_check_fails()
         {
-            var sut = new ScopedRights(Setup.ARandomScope, new short[] { ScopedRights.GlobalAdmin });
+            var sut = new ScopedRights(Setup.ARandomScope, new int[] { ScopedRights.GlobalAdmin });
             
             sut.HasRight(23).Should().BeFalse();
         }
